Added functionality to forecast in Window down to flood inundation maps and through HMS in Linux.

+++ run_hms_hrrr.py
@@ -69,11 +69,11 @@ def run(simulation_length,
 
     
     # see if HRRR directory already exists and if so, remove it
-    if os.path.exists(hrrr_directory):
+    if os.path.exists(hrrr_directory) and download_met_data == True:
         shutil.rmtree(hrrr_directory)
         os.mkdir(hrrr_directory)
     else:
-        os.mkdir(hrrr_directory)
+        pass
 
     # download the HRRR precip and point the script to the download folder
     hrrr_directory = download(hrrr_directory, current_time_latency_two_hour, download_met_data)
@@ -109,6 +109,74 @@ def run(simulation_length,
             open_vortex_file.write(string_to_write)
             open_vortex_file.write('\n')
             string_to_write = '{0} -J-Xmx12g -Djava.library.path=%VORTEX_HOME%\\bin;%VORTEX_HOME%\\bin\gdal met_data_import.py "{1}" "{2}" "{3}" {4} "{5}"\n'.format(path_to_jython_install,hrrr_directory,hec_hms_clip_shp,vortex_dss_file_path,variables,met_forcing)
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.close()
+
+            # Open the subprocess without creating a new window
+            process = subprocess.Popen(vortex_file, shell=True)
+            stdout, stderr = process.communicate()
+    elif sys.platform == 'linux' or sys.platform == 'linux2':
+        # we're building the bat file that runs Vortex in Linux here
+        print("Running Vortex on Linux.\n")
+        vortex_file_name = "vortex_{0}{1}{2}{3}.sh".format(current_time_latency_two_hour.year, 
+                                                    current_time_latency_two_hour.strftime("%m"),
+                                                    current_time_latency_two_hour.strftime("%d"),
+                                                    current_time_latency_two_hour.strftime("%H"))
+        vortex_file = os.path.join(hrrr_directory,vortex_file_name)
+        with open(vortex_file, "w") as open_vortex_file:
+            string_to_write = 'VORTEX_HOME="{0}"\n'.format(str(path_to_vortex_install))
+            open_vortex_file.write(string_to_write)
+
+            string_to_write = 'export PATH="$VORTEX_HOME/bin:$VORTEX_HOME/bin/gdal:$PATH"'
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            string_to_write = 'export GDAL_DATA="$VORTEX_HOME/bin/gdal/gdal-data"' 
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            string_to_write = 'export PROJ_LIB="$VORTEX_HOME/bin/gdal/proj"'
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            string_to_write = 'JAVA_VORTEX_PATH="$VORTEX_HOME/jre/bin/java"' 
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            string_to_write = 'JAVA_VORTEX_LIB_PATH="$VORTEX_HOME/bin:$VORTEX_HOME/bin/gdal"'
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            string_to_write = 'export CLASSPATH="{0}:$VORTEX_HOME/lib/*"'.format(path_to_jython_install)
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            LD_LIBRARY_PATH = os.environ.get('LD_LIBRARY_PATH')
+            if LD_LIBRARY_PATH==None:
+                string_to_write = 'export LD_LIBRARY_PATH="$VORTEX_HOME/bin/gdal"'.format(LD_LIBRARY_PATH)
+            else:
+                string_to_write = 'export LD_LIBRARY_PATH="$VORTEX_HOME/bin/gdal:${0}"'.format(LD_LIBRARY_PATH)
+            open_vortex_file.write(string_to_write)
+            open_vortex_file.write('\n')
+
+            # string_to_write = 'ADD_OPENS="java.desktop/sun.awt.shell=ALL-UNNAMED"'
+            # open_vortex_file.write(string_to_write)
+            # open_vortex_file.write('\n')
+
+            # string_to_write = 'MAIN_CLASS="mil.army.usace.hec.vortex"'
+            # open_vortex_file.write(string_to_write)
+            # open_vortex_file.write('\n')
+
+            # string_to_write = 'APP_ARG="$1"'
+            # open_vortex_file.write(string_to_write)
+            # open_vortex_file.write('\n')
+
+            # string_to_write = 'export JYTHON_OPTS="-Xmx12g -Djava.library.path=$VORTEX_HOME/bin:$VORTEX_HOME/bin/gdal"'
+            # open_vortex_file.write(string_to_write)
+            # open_vortex_file.write('\n')
+
+            # string_to_write = '$JAVA_VORTEX_PATH -Xmx20g -Djava.library.path="$LD_LIBRARY_PATH" org.python.util.jython met_data_import.py "{1}" "{2}" "{3}" "{4}" "{5}"\n'.format(path_to_vortex_install,hrrr_directory,hec_hms_clip_shp,vortex_dss_file_path,variables,met_forcing,path_to_jython_install)
+            string_to_write = '$JAVA_VORTEX_PATH -Xmx20g -Djava.library.path="$JAVA_VORTEX_LIB_PATH" org.python.util.jython met_data_import.py "{1}" "{2}" "{3}" "{4}" "{5}"\n'.format(path_to_vortex_install,hrrr_directory,hec_hms_clip_shp,vortex_dss_file_path,variables,met_forcing,path_to_jython_install)
             open_vortex_file.write(string_to_write)
             open_vortex_file.close()
